Move _cache(), _get_cached() to _DeployableEntity
