+++ client/verta/verta/_tracking/deployable_entity.py
@@ -5,7 +5,10 @@ from __future__ import print_function
 import copy
 import glob
 import os
+import shutil
 import sys
+import tarfile
+import tempfile
 import zipfile
 
 import requests
@@ -25,6 +28,13 @@ from ..external import six
 # location in DeploymentService model container
 _CUSTOM_MODULES_DIR = os.environ.get('VERTA_CUSTOM_MODULES_DIR', "/app/custom_modules/")
 
+# for caching files
+_CACHE_DIR = os.path.join(
+    os.path.expanduser("~"),
+    ".verta",
+    "cache",
+)
+
 
 class _DeployableEntity(_ModelDBEntity):
     @property
@@ -34,6 +44,92 @@ class _DeployableEntity(_ModelDBEntity):
             self._conn.socket,
             self.id,
         )
+
+    def _cache_file(self, filename, contents):
+        """
+        Caches `contents` to `filename` within ``_CACHE_DIR``.
+
+        If `contents` represents a ZIP file, then it will be unzipped, and the path to the target
+        directory will be returned.
+
+        Parameters
+        ----------
+        filename : str
+            Filename within ``_CACHE_DIR`` to write to.
+        contents : bytes
+            Contents to be cached.
+
+        Returns
+        -------
+        str
+            Full path to cached contents.
+
+        """
+        # write contents to temporary file
+        with tempfile.NamedTemporaryFile(delete=False) as tempf:
+            tempf.write(contents)
+            tempf.flush()  # flush object buffer
+            os.fsync(tempf.fileno())  # flush OS buffer
+
+        name, extension = os.path.splitext(filename)
+        if extension == '.zip':
+            temp_path = tempfile.mkdtemp()
+
+            with zipfile.ZipFile(tempf.name, 'r') as zipf:
+                zipf.extractall(temp_path)
+            os.remove(tempf.name)
+        elif extension == '.tgz':
+            temp_path = tempfile.mkdtemp()
+
+            with tarfile.open(tempf.name, 'r:gz') as tarf:
+                tarf.extractall(temp_path)
+            os.remove(tempf.name)
+        elif extension == '.tar':
+            temp_path = tempfile.mkdtemp()
+
+            with tarfile.open(tempf.name, 'r') as tarf:
+                tarf.extractall(temp_path)
+            os.remove(tempf.name)
+        elif extension == '.gz' and os.path.splitext(name)[1] == '.tar':
+            name = os.path.splitext(name)[0]
+
+            temp_path = tempfile.mkdtemp()
+
+            with tarfile.open(tempf.name, 'r:gz') as tarf:
+                tarf.extractall(temp_path)
+            os.remove(tempf.name)
+        else:
+            name = filename
+            temp_path = tempf.name
+
+        path = os.path.join(_CACHE_DIR, name)
+
+        # create intermediate dirs
+        try:
+            os.makedirs(os.path.dirname(path))
+        except OSError:  # already exists
+            pass
+
+        # move written contents to cache location
+        shutil.move(temp_path, path)
+
+        return path
+
+    def _get_cached_file(self, filename):
+        name, extension = os.path.splitext(filename)
+        if extension == '.zip':
+            pass
+        elif extension == '.tgz':
+            pass
+        elif extension == '.tar':
+            pass
+        elif extension == '.gz' and os.path.splitext(name)[1] == '.tar':
+            name = os.path.splitext(name)[0]
+        else:
+            name = filename
+
+        path = os.path.join(_CACHE_DIR, name)
+        return path if os.path.exists(path) else None
 
     def _custom_modules_as_artifact(self, paths=None):
         if isinstance(paths, six.string_types):
